+++ src/presentation/controller/signup.spec.ts
@@ -1,8 +1,6 @@
 import { SignUpController } from './signup';
-import { MissingParamError } from '../errors/missing-param-error';
-import { InvalidParamError } from '../errors/invalid-param-error';
 import { EmailValidator } from '../protocols/email-validator.interface';
-import { ServerError } from '../errors/server-error';
+import { MissingParamError, InvalidParamError, ServerError } from '../errors';
 
 interface SutTypes {
   sut: SignUpController
